requirements.txt
anthropic
openai
instructor
pydantic
pandas
psycopg2-binary
python-dotenv
timescale-vector